% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lat-lng.R
\docType{data}
\name{example-lat-lng}
\alias{example-lat-lng}
\alias{lat_abby_rd}
\alias{lng_abby_rd}
\alias{lat_cambridge}
\alias{lng_cambridge}
\alias{lat_liverpool}
\alias{lng_liverpool}
\alias{lat_london}
\alias{lng_london}
\alias{lat_oxford}
\alias{lng_oxford}
\title{Example Longitude and Latitude for practice with ukpolice}
\format{An object of class \code{numeric} of length 1.}
\usage{
lat_abby_rd

lng_abby_rd

lat_cambridge

lng_cambridge

lat_liverpool

lng_liverpool

lat_london

lng_london

lat_oxford

lng_oxford
}
\description{
Example Longitude and Latitude for practice with ukpolice
}
\examples{
\dontrun{
ukp_crime(lat = lat_london, lng = lng_london)
}
}
\keyword{datasets}
