% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neighbourhoods.R
\name{get_neighbourhoods}
\alias{get_neighbourhoods}
\title{get_neighbourhoods}
\usage{
get_neighbourhoods(neighbourhood)
}
\arguments{
\item{neighbourhood}{a text string of a neighbourhood in the UK}
}
\value{
tibble with columns id and name. id is a Police force specific team identifier, note that this identifier is not unique and may also be used by a different force. Name is the name for the neighbourhood
}
\description{
This call retuns the list of neighbourhoods for a force, https://data.police.uk/docs/method/neighbourhoods/
}
\examples{

\dontrun{
library(ukpolice)
get_neighbourhoods("leicestershire")

}

}

