% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_constructors.R
\name{print.ukpolice_api}
\alias{print.ukpolice_api}
\title{print method for ukpolice_api}
\usage{
\method{print}{ukpolice_api}(x, ...)
}
\arguments{
\item{x}{ukpolice_api object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a special print method lifted from https://cran.r-project.org/web/packages/httr/vignettes/api-packages.html
}
\description{
print method for ukpolice_api
}

