% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime-category.R
\name{ukc_crime_category}
\alias{ukc_crime_category}
\title{Crime categories}
\usage{
ukc_crime_category()
}
\value{
A \code{tibble} with all available categories of crime.
}
\description{
Crime categories
}
