% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime-at-location.R
\name{ukc_crime_poly}
\alias{ukc_crime_poly}
\title{Extract crime areas within a polygon}
\usage{
ukc_crime_poly(poly_df, date = NULL, ...)
}
\arguments{
\item{poly_df}{dataframe containing the lat/lng pairs which define the
boundary of the custom area. If a custom area contains more than 10,000
crimes, the API will return a 503 status code. ukc_crime_poly converts the
dataframe into lat/lng pairs, separated by colons:
\code{lat},\code{lng}:\code{lat},\code{lng}:\code{lat},\code{lng}. The first and last coordinates need
not be the same — they will be joined by a straight line once the request
is made.}

\item{date, }{Optional. (YYY-MM), limit results to a specific month. The
latest month will be shown by default. e.g. date = "2013-01"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extract crime areas within a polygon
}
\note{
further documentation here:
\url{https://data.police.uk/docs/method/crime-street/}
}
\examples{

library(ukpolice)

# with 3 points
poly_df_3 = data.frame(lat = c(52.268, 52.794, 52.130),
                       long = c(0.543, 0.238, 0.478))

ukc_data_poly_3 <- ukc_crime_poly(poly_df_3)
head(ukc_data_poly_3)

# with 4 points
poly_df_4 = data.frame(lat = c(52.268, 52.794, 52.130, 52.000),
                       long = c(0.543,  0.238,  0.478,  0.400))
ukc_data_poly_4 <- ukc_crime_poly(poly_df = poly_df_4)

head(ukc_data_poly_4)

}
