% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last-update.R
\name{ukc_last_update}
\alias{ukc_last_update}
\title{Latest crime update}
\usage{
ukc_last_update()
}
\description{
Returns the latest month crime data was updated for. The date is in standard
ISO format but the actual day is not relevant.
}
