% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood-location.R
\name{ukc_neighbourhood_location}
\alias{ukc_neighbourhood_location}
\title{Neighbourhood Location}
\usage{
ukc_neighbourhood_location(lat, lng)
}
\arguments{
\item{lat}{Latitude}

\item{lng}{Longitude}
}
\value{
The police force and neighbourhood code of the given coordinates.
}
\description{
Find the neighbourhood policing team responsible for a given area, by
a set of coordinates.
}
\examples{
\dontrun{
find <- ukc_neighbourhood_location(lat = 51.500617, lng = -0.124629)
}

}
\seealso{
\code{\link[=ukc_neighbourhood_boundary]{ukc_neighbourhood_boundary()}}

\code{\link[=ukc_neighbourhoods]{ukc_neighbourhoods()}}
}
