% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhoods.R
\name{ukc_neighbourhoods}
\alias{ukc_neighbourhoods}
\title{Policing Neighbourhoods}
\usage{
ukc_neighbourhoods(force)
}
\arguments{
\item{force}{A string containing the name of the police force to return
neighbourhoods for. Must be specified, and is not case sensitive.}
}
\value{
A \code{tibble} with data for neighbourhoods within the area of the
given police force.
}
\description{
All the neighbourhoods within a given police force area.
}
\examples{
\dontrun{
places <- ukc_neighbourhoods("dorset")
}

}
\seealso{
\code{\link[=ukc_neighbourhood_boundary]{ukc_neighbourhood_boundary()}}

\code{\link[=ukc_neighbourhood_specific]{ukc_neighbourhood_specific()}}

\code{\link[=ukc_neighbourhood_location]{ukc_neighbourhood_location()}}
}
