% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime-outcome.R
\name{ukc_street_crime_outcome}
\alias{ukc_street_crime_outcome}
\title{Street level crime outcomes}
\usage{
ukc_street_crime_outcome(lat, lng, location, date = NULL)
}
\arguments{
\item{lat}{Latitude. Accepts a single value or a vector of values to create
a custom polygon.}

\item{lng}{Longitude. Accepts a single value or a vector of values to create
a custom polygon.}

\item{location}{If specified, \code{lat} and \code{lng} are ignored. Location IDs are
available through other methods including \code{\link[=ukc_street_crime]{ukc_street_crime()}}.}

\item{date}{The year and month in "YYYY-MM" form. If \code{NULL}, latest
available month will be returned.}
}
\value{
A tibble with details of street crime outcomes.
}
\description{
Street level crime outcomes
}
\details{
If specified, \code{lat} and \code{lng} must be the same length. \code{location}
or both \code{lat} and \code{lng} must be specified.
}
