% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime.R
\name{ukp_crime_unlist}
\alias{ukp_crime_unlist}
\title{ukp_crime_unlist}
\usage{
ukp_crime_unlist(result_content)
}
\arguments{
\item{result_content}{a result from the ukp_api}
}
\value{
data frame
}
\description{
a utility function to clean and unlist the data extracted from the crime data
}

