% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forces.R
\name{ukp_forces}
\alias{ukp_forces}
\title{List Information about a specific force in the UK}
\usage{
ukp_forces(force)
}
\description{
List specific information for a given police force in the UK.
}
\details{
https://data.police.uk/docs/method/force/
returns this information
}
