% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_crime.R
\name{ukp_geo_chull}
\alias{ukp_geo_chull}
\title{ukp_geo_chull}
\usage{
ukp_geo_chull(data, long, lat)
}
\arguments{
\item{data}{a dataframe}

\item{long}{longitude}

\item{lat}{latitude}
}
\value{
a
}
\description{
compute the convex hull of some lon/lat points
}
\examples{

\dontrun{

library(maxcovr)

# identify the polygon you want to draw
poly_string <- ukp_geo_chull(data = york,
                             long = long,
                             lat = lat)

ukp_crime_poly()


}

}

