% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forces.R
\name{ukp_list_forces}
\alias{ukp_list_forces}
\title{List Police Forces in the UK}
\usage{
ukp_list_forces()
}
\value{
table with columns \code{id} and \code{name}.\code{id} is a Unique force identifier,
\code{name} is the force name.
}
\description{
This is an empty function called as \code{ukp_list_forces()}. This lists all the
police forces available via the UK police API. The unique force identifiers
can be used in requests for force-specific data. Such as \code{\link[=ukp_force]{ukp_force()}} and
\code{\link[=ukp_force_people]{ukp_force_people()}}.
}
\examples{
ukp_list_forces()
}
