% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhoods.R
\name{ukp_neighbourhood}
\alias{ukp_neighbourhood}
\title{API calls around the neighbourhoods}
\usage{
ukp_neighbourhood(neighbourhood)
}
\arguments{
\item{neighbourhood}{a text string of a neighbourhood in the UK}
}
\value{
tibble with columns id and name. id is a Police force specific team identifier, note that this identifier is not unique and may also be used by a different force. Name is the name for the neighbourhood
}
\description{
ukp_neighbourhoods
}
\details{
This call retuns the list of neighbourhoods for a force, https://data.police.uk/docs/method/neighbourhoods/
}
\examples{

\dontrun{
library(ukpolice)
ukp_neighbourhood("leicestershire")

}

}

