% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhoods.R
\name{ukp_neighbourhood_location}
\alias{ukp_neighbourhood_location}
\title{Find the assoiate neighbourhood code with a long/lat string}
\usage{
ukp_neighbourhood_location(lat_long, ...)
}
\arguments{
\item{lat_long}{a character string of latitude and longitude}

\item{...}{further arguments passed to or from other methods. For example,
verbose option can be added with ukp_api("call", config = httr::verbose()).
See more in ?httr::GET documentation
(\url{https://cran.r-project.org/web/packages/httr/}) and
(\url{https://cran.r-project.org/web/packages/httr/vignettes/quickstart.html}).}
}
\value{
a tibble
}
\description{
\code{ukp_neighbourhood_location} takes a longitude/latitude string and finds the
associated neighbourhood code associated with that. Read more here
https://data.police.uk/docs/method/neighbourhood-locate/
}
