% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_crime.R
\name{ukp_poly_paste}
\alias{ukp_poly_paste}
\title{ukp_poly_paste}
\usage{
ukp_poly_paste(data, long, lat)
}
\arguments{
\item{data}{a dataframe}

\item{long}{character}

\item{lat}{character}
}
\value{
character vector: "lat_1,long_1:lat_2,long_2:...,lat_n,long_n"
}
\description{
Takes a dataframe oif longitude and latitude and pastes them into the correct
    format required for the ukpolice API
}
\examples{

\dontrun{

library(maxcovr)

ukp_poly_paste_2(york, "long", "lat")

}

}

