% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/york.R
\docType{data}
\name{york}
\alias{york}
\title{Listed buildings provided by the City of York Council, made available
\href{https://data.gov.uk/dataset/9417652f-3901-4a16-8368-222deefb36cd/listed-buildings}{here}.
This data contains public sector information licensed under the Open
Government Licence v3.0:
\url{https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/}.}
\format{A data frame with seven variables: \code{long}, \code{lat}, \code{desig_id},
\code{bldg_name}, \code{grade}, \code{status_date}, and \code{esri_oid}.

\describe{
\item{\code{long}}{longitude of the building}
\item{\code{lat}}{latitude of the building}
\item{\code{desig_id}}{ID related to a feature that is not yet known to me}
\item{\code{bldg_name}}{name of the building}
\item{\code{grade}}{one of the three (I, II, III) categories of listed buildings}
\item{\code{status_date}}{date the status was made}
\item{\code{esri_oid}}{ID related to a feature that is not yet known to me}
}

For further details, see \url{https://www.york.gov.uk/info/20215/conservation_and_listed_buildings/1346/listed_buildings} and \url{https://data.gov.uk/dataset/9417652f-3901-4a16-8368-222deefb36cd/listed-buildings}}
\usage{
york
}
\description{
Listed buildings provided by the City of York Council, made available
\href{https://data.gov.uk/dataset/9417652f-3901-4a16-8368-222deefb36cd/listed-buildings}{here}.
This data contains public sector information licensed under the Open
Government Licence v3.0:
\url{https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/}.
}
\keyword{datasets}
