% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_constructors.R
\name{ukp_api}
\alias{ukp_api}
\title{ukpolice_api}
\usage{
ukp_api(path, ...)
}
\arguments{
\item{path}{character}

\item{...}{further arguments passed to or from other methods. For example, verbose option can be added with ukp_api("call", config = httr::verbose())}
}
\value{
a ukpolice_api object
}
\description{
ukpolice_api
}

