% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_crime.R
\name{ukp_crime_unlist}
\alias{ukp_crime_unlist}
\title{Unlist and clean crime data}
\usage{
ukp_crime_unlist(result_content)
}
\arguments{
\item{result_content}{a result from the ukp_api}
}
\value{
data.frame
}
\description{
\code{ukp_crime_unlist} is a utility function to clean and unlist the data extracted from the crime data.
}

