% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_crime.R
\name{ukp_geo_chull}
\alias{ukp_geo_chull}
\title{Compute convex hull of long/lat points}
\usage{
ukp_geo_chull(data, long, lat)
}
\arguments{
\item{data}{a dataframe}

\item{long}{longitude, a character naming the long name}

\item{lat}{latitude, a character naming the lat name}
}
\value{
a dataframe with the positions for the polygon.
}
\description{
\code{ukp_geo_chull} computes the convex hull of some lon/lat points, returning a dataframe of the polygon position
}
\examples{

ukp_geo_chull(york,
              "long",
              "lat")
}
