% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhoods.R
\name{ukp_neighbourhood_team}
\alias{ukp_neighbourhood_team}
\title{ukp_neighbourhood_team}
\usage{
ukp_neighbourhood_team(lat_long, ...)
}
\arguments{
\item{lat_long}{a character string of latitude and longitude}

\item{...}{further arguments passed to or from other methods. For example, verbose option can be added with ukp_api("call", config = httr::verbose())}
}
\value{
a tibble
}
\description{
\code{ukp_neighbourhood_team} takes a longitude/latitude string and finds the associated neighbourhood code associated with that. Read more here https://data.police.uk/docs/method/neighbourhood-locate/
}

