% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_crime.R
\name{ukp_poly_paste}
\alias{ukp_poly_paste}
\title{Format long/lat numbers into the right string}
\usage{
ukp_poly_paste(data, long, lat)
}
\arguments{
\item{data}{a dataframe}

\item{long}{character}

\item{lat}{character}
}
\value{
character vector: "lat_1,long_1:lat_2,long_2:...,lat_n,long_n"
}
\description{
\code{ukp_poly_paste} takes a dataframe of longitude and latitude and pastes
them into the correct format required for the ukpolice API
}
\examples{

\dontrun{
ukp_poly_paste(york, "long", "lat")
}
}
