% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_and_search.R
\name{ukp_stop_search}
\alias{ukp_stop_search}
\alias{ukp_stop_search_poly}
\title{Find stop and search at street level}
\usage{
ukp_stop_search(lat, lng, date = NULL, ...)

ukp_stop_search_poly(poly_df, date = NULL, ...)
}
\arguments{
\item{lat}{latitude of the requested crime area}

\item{lng, }{longitude of the requested crime area}

\item{date, }{Optional. (YYY-MM), limit results to a specific month. The
latest month will be shown by default. e.g. date = "2013-01"}

\item{...}{further arguments passed to or from other methods. For example,
verbose option can be added with
\code{ukp_api("call", config = httr::verbose())}. See more in \code{?httr::GET}
documentation
(\url{https://cran.r-project.org/web/packages/httr/}) and
(\url{https://cran.r-project.org/web/packages/httr/vignettes/quickstart.html}).}

\item{poly_df}{dataframe containing the lat/lng pairs which define the
boundary of the custom area. If a custom area contains more than 10,000
crimes, the API will return a 503 status code. ukp_crime_poly converts the
dataframe into lat/lng pairs, separated by colons:
\code{lat},\code{lng}:\code{lat},\code{lng}:\code{lat},\code{lng}. The first and last coordinates need
not be the same — they will be joined by a straight line once the request
is made.}
}
\value{
a tibble with the columns:
\itemize{
\item type:	Whether this was a 'Person search', a 'Vehicle search', or a
'Person and Vehicle search'
\item involved_person:	Whether a person was searched in this incident
(derived from type; true if anything but 'Vehicle search')
\item datetime:	When the stop and search took place. Note that some forces
only provide dates for their stop and searches, so you might see a
disproportionate number of incidents occurring at midnight.
\item operation: (not always present in data)	Whether this stop and search
was part of a policing operation
\item operation_name: (not always present in data) name of the operation this
stop and search was part of
\item location:	Approximate location of the incident
\item latitude:	Latitude
\item longitude:	Longitude
\item street:	The approximate location of the incident
\item id:	Unique identifier for the location
\item name:	Human-readable summary of the location
\item gender:	Human-readable gender of the person stopped, if applicable
and provided
\item age_range:	The age range of the person stopped at the time the stop
occurred
\item self_defined_ethnicity:	The self-defined ethnicity of the person stopped
\item officer_defined_ethnicity:	The officer-defined ethnicity of the person
stopped
\item legislation:	The power used to carry out the stop and search
\item object_of_search:	The reason the stop and search was carried out
\item outcome:	The outcome of the stop. false if nothing was found, an empty
string if no outcome was provided.
\item outcome_linked_to_object_of_search:	Whether the outcome was related to
the reason the stop and search was carried out, as a boolean value
(or null if not provided)
\item removal_of_more_than_outer_clothing:	Whether the person searched had
more than their outer clothing removed, as a boolean value (or null if
not provided)
}
}
\description{
Crimes at street-level; either within a 1 mile radius of a single point
(\code{ukp_stop_search}), or within a custom area (\code{ukp_stop_search_poly}).
The street-level crimes returned in the API are only an approximation of
where the actual crimes occurred, they are not the exact locations. See
the about page (\url{https://data.police.uk/about/#location-anonymisation})
for more information about location anonymisation. Note that crime
levels may appear lower in Scotland, as only the British Transport
Police provide this data.
}
\note{
The API will return a 400 status code in response to a GET request
longer than 4094 characters. For submitting particularly complex poly
parameters, consider using POST instead.

more documentation here:
\url{https://data.police.uk/docs/method/stops-street/}
}
\examples{

ukp_stop_search_data <- ukp_stop_search(lat = 52.629729,
                                        lng = -1.131592)

head(ukp_stop_search_data)


# with 3 points
poly_df_3 = data.frame(lat = c(52.268, 52.794, 52.130),
                       long = c(0.543, 0.238, 0.478))

ukp_data_poly_3 <- ukp_stop_search_poly(poly_df_3)
head(ukp_data_poly_3)

# with 4 points
poly_df_4 = data.frame(lat = c(52.268, 52.794, 52.130, 52.000),
                       long = c(0.543,  0.238,  0.478,  0.400))
ukp_data_poly_4 <- ukp_stop_search_poly(poly_df = poly_df_4)

head(ukp_data_poly_4)

}
