% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukpolice-package.R
\docType{package}
\name{ukpolice}
\alias{ukpolice-package}
\title{ukpolice: Download data on police forces and reported crime in England and Wales}
\description{
ukpolice: Download data on police forces and reported crime in England and Wales
}
