% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukpolice-package.R
\name{ukpolice}
\alias{ukpolice}
\title{ukpolice}
\description{
A package to retrieve data from the UK police data repository: ukpolice is
an R package that facilitates retrieving data from the
\href{https://data.police.uk/}{UK police database.}. The data provided by
the API contains public sector information licensed under the
\href{http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/}{Open Government Licence v3.0.}.
}
